Functionality added, refactoring and cleanup
added functionality to do reservations from the favourites/admin-resourcepage. Non-admin reservations are checked for min/max duration restrictions, Added functionality to unselect by clicking the first selected slot again. Once a slot is selected the popover shows the min/max times for the resource, if selection is under min duration it wont react. This is a WIP