Added favorites page
Added a new favorites page, nonAdmin users are now directed to this page instead of the old admin-resources page.
Added new reducer for the favorites and all the necessary uiActions/ActionTypes.
Added tests to all of the files and added new attributes to the fixtures/Resource that are needed for the tests.