Resources reflect userpermissions
Page now displays resources based on if the unit id is in user.staffStatus.isManagerFor[] array. If user is marked as superuser in respa then the page will display all resources from every unit. If the user is marked as a unitmanager, unitadministrator or unit group administrator the unit ids are all grouped in the isManagerFor array.